Admin views are now Vue files
